+++ antlr/src/test/kotlin/com/strumenta/kolasu/antlr/mapping/ParseTreeToASTTransformerTest.kt
@@ -1,12 +1,12 @@
 package com.strumenta.kolasu.antlr.mapping
 
 import com.strumenta.kolasu.antlr.parsing.withParseTreeNode
+import com.strumenta.kolasu.ast.Range
 import com.strumenta.kolasu.model.GenericErrorNode
 import com.strumenta.kolasu.model.Named
 import com.strumenta.kolasu.model.Node
 import com.strumenta.kolasu.model.NodeLike
 import com.strumenta.kolasu.model.PossiblyNamed
-import com.strumenta.kolasu.model.Range
 import com.strumenta.kolasu.model.ReferenceByName
 import com.strumenta.kolasu.model.children
 import com.strumenta.kolasu.model.hasValidParents
