+++ core/src/main/kotlin/com/strumenta/kolasu/model/Destinations.kt
@@ -1,6 +1,7 @@
 package com.strumenta.kolasu.model
 
 import com.strumenta.kolasu.ast.Destination
+import com.strumenta.kolasu.ast.Range
 
 data class CompositeDestination(
     val elements: List<Destination>,
