+++ core/src/main/kotlin/com/strumenta/kolasu/model/Errors.kt
@@ -1,5 +1,6 @@
 package com.strumenta.kolasu.model
 
+import com.strumenta.kolasu.ast.Range
 import com.strumenta.kolasu.traversing.walkDescendants
 
 /**
