+++ core/src/main/kotlin/com/strumenta/kolasu/transformation/Transformation.kt
@@ -1,11 +1,11 @@
 package com.strumenta.kolasu.transformation
 
+import com.strumenta.kolasu.ast.Range
 import com.strumenta.kolasu.model.GenericErrorNode
 import com.strumenta.kolasu.model.NodeLike
 import com.strumenta.kolasu.model.NodeOrigin
 import com.strumenta.kolasu.model.Origin
 import com.strumenta.kolasu.model.PropertyTypeDescription
-import com.strumenta.kolasu.model.Range
 import com.strumenta.kolasu.model.children
 import com.strumenta.kolasu.model.processProperties
 import com.strumenta.kolasu.model.withOrigin
