+++ core/src/main/kotlin/com/strumenta/kolasu/traversing/ByRange.kt
@@ -2,8 +2,9 @@
 
 package com.strumenta.kolasu.traversing
 
+import com.strumenta.kolasu.ast.Range
 import com.strumenta.kolasu.model.NodeLike
-import com.strumenta.kolasu.model.Range
+import com.strumenta.kolasu.model.contains
 
 /**
  * @param range the range where to search for nodes
