+++ core/src/test/kotlin/com/strumenta/kolasu/model/RangeTest.kt
@@ -1,5 +1,8 @@
 package com.strumenta.kolasu.model
 
+import com.strumenta.kolasu.ast.Point
+import com.strumenta.kolasu.ast.Range
+import com.strumenta.kolasu.ast.START_POINT
 import kotlin.test.assertEquals
 import kotlin.test.assertFails
 import kotlin.test.assertFalse
