+++ core/src/test/kotlin/com/strumenta/kolasu/transformation/ASTTransformerTest.kt
@@ -1,9 +1,9 @@
 package com.strumenta.kolasu.transformation
 
+import com.strumenta.kolasu.ast.Range
 import com.strumenta.kolasu.model.Node
 import com.strumenta.kolasu.model.NodeLike
 import com.strumenta.kolasu.model.NodeOrigin
-import com.strumenta.kolasu.model.Range
 import com.strumenta.kolasu.model.hasValidParents
 import com.strumenta.kolasu.model.withOrigin
 import com.strumenta.kolasu.testing.assertASTsAreEqual
