+++ emf/src/main/kotlin/com/strumenta/kolasu/emf/Metamodel.kt
@@ -1,5 +1,6 @@
 package com.strumenta.kolasu.emf
 
+import com.strumenta.kolasu.ast.Range
 import com.strumenta.kolasu.model.EntityDeclaration
 import com.strumenta.kolasu.model.ErrorNode
 import com.strumenta.kolasu.model.Expression
@@ -9,7 +10,6 @@ import com.strumenta.kolasu.model.Node
 import com.strumenta.kolasu.model.NodeLike
 import com.strumenta.kolasu.model.PlaceholderElement
 import com.strumenta.kolasu.model.PossiblyNamed
-import com.strumenta.kolasu.model.Range
 import com.strumenta.kolasu.model.ReferenceByName
 import com.strumenta.kolasu.model.Statement
 import com.strumenta.kolasu.transformation.GenericNode
