+++ emf/src/test/kotlin/com/strumenta/kolasu/emf/ResultTest.kt
@@ -1,9 +1,9 @@
 package com.strumenta.kolasu.emf
 
+import com.strumenta.kolasu.ast.Point
+import com.strumenta.kolasu.ast.Range
 import com.strumenta.kolasu.emf.serialization.JsonGenerator
 import com.strumenta.kolasu.model.NodeLike
-import com.strumenta.kolasu.model.Point
-import com.strumenta.kolasu.model.Range
 import com.strumenta.kolasu.model.range
 import com.strumenta.kolasu.validation.Issue
 import com.strumenta.kolasu.validation.IssueSeverity
