+++ emf/src/test/kotlin/com/strumenta/kolasu/emf/rpgast/DataDefinitions.kt
@@ -3,9 +3,9 @@ package com.strumenta.kolasu.emf.rpgast
 import com.smeup.rpgparser.parsing.ast.Expression
 import com.strumenta.kolasu.ast.Derived
 import com.strumenta.kolasu.ast.Link
+import com.strumenta.kolasu.ast.Range
 import com.strumenta.kolasu.model.Named
 import com.strumenta.kolasu.model.Node
-import com.strumenta.kolasu.model.Range
 import java.math.BigDecimal
 import java.util.Locale
