+++ emf/src/test/kotlin/com/strumenta/kolasu/emf/rpgast/Misc.kt
@@ -1,7 +1,7 @@
 package com.strumenta.kolasu.emf.rpgast
 
+import com.strumenta.kolasu.ast.Range
 import com.strumenta.kolasu.model.NodeLike
-import com.strumenta.kolasu.model.Range
 
 data class ToAstConfiguration(
     val considerRange: Boolean = true,
