+++ emf/src/test/kotlin/com/strumenta/kolasu/emf/rpgast/PredefinedIndicatorExpr.kt
@@ -1,6 +1,6 @@
 package com.smeup.rpgparser.parsing.ast
 
-import com.strumenta.kolasu.model.Range
+import com.strumenta.kolasu.ast.Range
 
 // *IN01..*IN99
 data class PredefinedIndicatorExpr(
