+++ serialization/src/test/kotlin/JsonDeserializerTest.kt
@@ -1,8 +1,8 @@
 package com.strumenta.kolasu.serialization
 
 import com.google.gson.stream.JsonWriter
-import com.strumenta.kolasu.model.Point
-import com.strumenta.kolasu.model.Range
+import com.strumenta.kolasu.ast.Point
+import com.strumenta.kolasu.ast.Range
 import com.strumenta.kolasu.validation.Issue
 import com.strumenta.kolasu.validation.IssueType
 import com.strumenta.kolasu.validation.Result
